Consolidated all actions of creating a new App.config into one class. This makes it easier to read what is going on and cleaner.
